Add mapping to open diagnostics in float
Also tweaked the quick fix lsp mapping from <leader>q (for quickfix) to
<leader>f (for fix).

This mirrors the code navigation mapping to show floats which is gf. All
the code navigation mappings are g<x>. gd could have been a meaningful
mapping (d for diagnostics) but it's already used by "go to definition".

Moreover, having the same key for "show something about this line" and
"automatically do something about this line" is convenient for my muscle
memory, and since diagnostics usually relate to something that might
need to be fixed, having the mapping be gf makes some kind of sense.