Better persistence behavior
Made it behave like auto-session does: when starting nvim with no
file, like `nvim`, auto-load the session associated with the cwd.

This was alrady happening. But crucially, what auto-session does is that
it doesn't save sessions when one _does_ start nvim with a file. This
part was missing from my config.

It is important because if I have a session with several open windows,
then quit, and then just so happen to run something like
`nvim ~/some_path/notes.txt`, I don't want `notes.txt` to take over the
previous session simply because I happened to run nvim from the same
directory.

So persistence now gets stopped when opening nvim with a file, to avoid
the above.